restructure classes
