+++ app/src/main/java/com/vgscollect/androiddemo/usecase/payments/checkout/collect/CollectCheckoutFormActivity.kt
@@ -1,4 +1,4 @@
-package com.vgscollect.androiddemo
+package com.vgscollect.androiddemo.usecase.payments.checkout.collect
 
 import android.os.Bundle
 import android.widget.Toast
@@ -9,10 +9,11 @@ import com.verygoodsecurity.vgscollect.core.VGSCollect
 import com.verygoodsecurity.vgscollect.core.VgsCollectResponseListener
 import com.verygoodsecurity.vgscollect.core.model.network.VGSRequest
 import com.verygoodsecurity.vgscollect.core.model.network.VGSResponse
+import com.vgscollect.androiddemo.R
 import kotlinx.android.synthetic.main.activity_collect_components.*
 import kotlinx.android.synthetic.main.activity_main.*
 
-class CollectActivity: AppCompatActivity() {
+class CollectCheckoutFormActivity: AppCompatActivity() {
 
     private val vgsForm: VGSCollect by lazy {
         VGSCollect.Builder(this, "tntpszqgikn")
@@ -44,8 +45,8 @@ class CollectActivity: AppCompatActivity() {
             override fun onResponse(response: VGSResponse?) {
                 when(response?.code) {
                     null -> return
-                    200 -> Toast.makeText(this@CollectActivity, "Card added!", Toast.LENGTH_SHORT).show()
-                    else -> Toast.makeText(this@CollectActivity, " $response", Toast.LENGTH_SHORT).show()
+                    200 -> Toast.makeText(this@CollectCheckoutFormActivity, "Card added!", Toast.LENGTH_SHORT).show()
+                    else -> Toast.makeText(this@CollectCheckoutFormActivity, " $response", Toast.LENGTH_SHORT).show()
                 }
 
             }
