+++ app/src/main/java/com/vgscollect/androiddemo/usecase/payments/demo/MaskAdapter.kt
@@ -1,4 +1,4 @@
-package com.vgscollect.androiddemo.vgs
+package com.vgscollect.androiddemo.usecase.payments.demo
 
 import com.verygoodsecurity.vgscollect.view.card.CardType
 import com.verygoodsecurity.vgscollect.view.card.formatter.CardMaskAdapter
